Modified Service, Controller and Repository classes

+++ PriceService/src/main/java/org/example/PriceService.java
@@ -4,6 +4,6 @@ import org.example.payload.PriceRequest;
 
 public interface PriceService {
 
-    int calculatePrice(PriceRequest priceRequest);
+    double calculatePrice(PriceRequest priceRequest);
 
 }
